Here's the code for the Fan of the Future project.
I based this code on existing templates by Espressif, the company behind the ESP32-CAM board and camera. On top of the code which already existed, I added the logic and some new code to convert the coordinates on the image into degrees for a servo motor to turn. Please feel free to improve.